Add img and examples
This change adds an image function, since the method for creating
a responsive image with the existing `view` function is a bit tedious:
you have to have a div inside with a background image set to the image
you want. This `img` function prevents the need for the extra div,
accomplishing it with an img with `object-fit: cover`

There are now two examples, one of an image and one of an iframe.